edgar-bridge/requirements.txt
edgartools>=3.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
